Extract depth conversion pipeline into engine/src/depth/depthConversionPipeline.ts
Moves handleDepthData, showDepthConversionUI, processDepthWithDefaults,
and processDepthWithParams out of main.ts as thin wrappers. Drops the
unused isAddFile parameter from processDepthWithDefaults (it was never
read in the original body).
